Loading the image to cloudinary

+++ src/modules/daybook/views/EntryView.vue
@@ -7,6 +7,13 @@
         <span class="mx-2 fs-4 fw-light">{{ yearDay }}</span>
       </div>
       <div>
+        <input
+          v-show="false"
+          ref="imageSelector"
+          @change="onImageSelected"
+          type="file"
+          accept="image/png, image/jpeg"
+        />
         <button
           v-if="entry?.id"
           @click="removeEntry"
@@ -15,7 +22,7 @@
           Borrar
           <i class="fa fa-trash-alt"></i>
         </button>
-        <button class="btn btn-primary">
+        <button @click="onSelectImage" class="btn btn-primary">
           Subir foto
           <i class="fa fa-upload"></i>
         </button>
@@ -28,7 +35,14 @@
   </template>
   <Fab icon="fa-save" @onClick="saveEntry" />
   <img
-    src="https://upload.wikimedia.org/wikipedia/commons/b/b2/JPEG_compression_Example.jpg"
+    v-if="entry?.picture && !localImage"
+    :src="entry.picture"
+    alt="entry-picture"
+    class="img-thumbnail"
+  />
+  <img
+    v-if="localImage"
+    :src="localImage"
     alt="entry-picture"
     class="img-thumbnail"
   />
@@ -39,10 +53,11 @@
 
   /** Third-Party */
   import { mapGetters, mapActions } from "vuex";
+  import Swal from "sweetalert2";
 
   /** Own */
   import { EntryType } from "../store/journal/types";
-  import { getDayMonthYear } from "../helpers";
+  import { getDayMonthYear, uploadImage } from "../helpers";
   import { EntryViewDataType } from "../types";
 
   export default defineComponent({
@@ -58,6 +73,8 @@
     data(): EntryViewDataType {
       return {
         entry: null,
+        localImage: null,
+        file: null,
       };
     },
     computed: {
@@ -109,6 +126,17 @@
         this.entry = entry;
       },
       async saveEntry() {
+        Swal.fire({
+          title: "Espere por favor",
+          allowOutsideClick: true,
+          backdrop: true,
+          showConfirmButton: false,
+        });
+        Swal.showLoading();
+
+        if (this.entry)
+          this.entry.picture = this.file ? await uploadImage(this.file) : null;
+
         if (this.entry?.id) {
           await this.updateEntry(this.entry);
         } else {
@@ -116,13 +144,58 @@
 
           this.$router.push({ name: "entry", params: { id: entryId } });
         }
+
+        Swal.fire("Guardado", "Entrada registrada con éxito", "success");
+        this.file = null;
       },
       async removeEntry() {
-        if (this.entry?.id) {
+        const { isConfirmed } = await Swal.fire({
+          title: "¿Estas seguro?",
+          text: "Una vez borrado, no se puede recuperar",
+          showDenyButton: true,
+          confirmButtonText: "Si, estoy seguro",
+        });
+
+        if (this.entry?.id && isConfirmed) {
+          Swal.fire({
+            title: "Espere por favor",
+            allowOutsideClick: true,
+            backdrop: true,
+          });
+          Swal.showLoading();
+
           const entryId: string | null = await this.deleteEntry(this.entry.id);
 
-          if (entryId) this.$router.push({ name: "no-entry" });
+          if (entryId) {
+            this.$router.push({ name: "no-entry" });
+            Swal.fire("Eliminado", "", "success");
+          } else {
+            Swal.fire(
+              "No se ha podido eliminar, intena nuevamente",
+              "",
+              "error"
+            );
+          }
         }
+      },
+      onImageSelected(event: Event) {
+        const { files } = event.target as any;
+
+        const file = files.length && files[0];
+
+        if (!file) {
+          this.file = null;
+          return;
+        }
+
+        this.file = file;
+
+        const fr = new FileReader();
+        fr.onload = () => (this.localImage = fr.result);
+        fr.readAsDataURL(file);
+      },
+      onSelectImage() {
+        (this.$refs.imageSelector as any).click();
       },
     },
     created() {
